Move chip tools
