+++ src/nexgen/beamlines/I24_Eiger_nxs.py
@@ -14,8 +14,8 @@ from .. import get_iso_timestamp, get_nexus_filename, log
 from ..nxs_write.NexusWriter import ScanReader, call_writers
 from ..nxs_write.NXclassWriters import write_NXdatetime, write_NXentry, write_NXnote
 from ..tools.VDS_tools import image_vds_writer
-from . import compute_goniometer, read_chip_map
 from .I24_Eiger_params import dset_links, eiger9M_params, goniometer_axes, source
+from .SSX_chip import compute_goniometer, read_chip_map
 
 # Define a logger object and a formatter
 logger = logging.getLogger("nexgen.I24")
